refactor: manager controllers refactored
