+++ backend/src/controllers/manager/assignEmployeeCourses/utils/validations.ts
@@ -1,13 +1,7 @@
-import { intersectionWith, isEmpty, uniqWith } from 'lodash';
+import { intersectionWith, isEmpty } from 'lodash';
 
 import { ICourseToAssign } from 'interfaces/ICourses/IQueryCourses';
 import { IClientCoursePopulated } from 'interfaces/Ientities/IclientCourses';
-
-const coursesToAssignComparator = (a: ICourseToAssign, b: ICourseToAssign): boolean =>
-  a.courseId === b.courseId;
-
-const removeCoursesToAssignDuplicates = (courses: ICourseToAssign[]): ICourseToAssign[] =>
-  uniqWith(courses, coursesToAssignComparator);
 
 const courseToCourseToAssignComparator = (a: IClientCoursePopulated, b: ICourseToAssign): boolean =>
   String(a.course._id) === String(b.courseId);
@@ -17,4 +11,4 @@ const isCoursesToAssignHaveDuplicates = (
   toAssign: ICourseToAssign[],
 ): boolean => !isEmpty(intersectionWith(courses, toAssign, courseToCourseToAssignComparator));
 
-export { removeCoursesToAssignDuplicates, isCoursesToAssignHaveDuplicates };
+export { isCoursesToAssignHaveDuplicates };
