+++ backend/src/controllers/manager/getPendingAssessments/utils/converters.ts
@@ -1,7 +1,6 @@
 import { IClientCoursePopulated } from 'interfaces/Ientities/IclientCourses';
 import { TAssessmentRequest } from 'interfaces/IResponse/IResponse';
-
-import { calculateTimeElapsed } from '../datetime/datetimeTypeConversions';
+import { calculateTimeElapsed } from 'utils/typeConversion/datetime/datetimeTypeConversions';
 
 const convertToAssessmentsRequests = (courses: IClientCoursePopulated[]): TAssessmentRequest[] =>
   courses.map((clCourse) => ({
