+++ backend/src/controllers/manager/index.ts
@@ -1,0 +1,5 @@
+export { default as assignEmployeeCourses } from './assignEmployeeCourses/assignEmployeeCourses';
+export { default as getEmployeeAvailableCourses } from './getEmployeeAvailableCourses/getEmployeeAvailableCourses';
+export { default as getEmployeeInfo } from './getEmployeeInfo/getEmployeeInfo';
+export { default as getEmployees } from './getEmployees/getEmployees';
+export { default as getPendingAssessments } from './getPendingAssessments/getPendingAssessments';
