+++ backend/src/utils/normaliser/employees.ts
@@ -1,13 +1,12 @@
 import { IClientCoursePopulated } from 'interfaces/Ientities/IclientCourses';
 import { TUserStackMember, TUserStackMemberShort } from 'interfaces/Ientities/IStackMember';
-import { IUser } from 'interfaces/Ientities/Iusers';
-import {
-  IEmployeeClientCourse,
-  IEmployeeInfo,
-  IEmployeeShortInfo,
-  TEmployeeCourse,
-} from 'interfaces/IResponse/IResponse';
-import { convertToTypeUnsafe } from 'utils/typeConversion/common';
+import { IEmployeeClientCourse, TEmployeeCourse } from 'interfaces/IResponse/IResponse';
+
+const mapEmployeeStack = (stack: TUserStackMember[]): TUserStackMemberShort[] =>
+  stack.map((stackMember) => ({
+    isPrimary: stackMember.isPrimary,
+    member: { name: stackMember.member.name },
+  }));
 
 const mapEmployeeClientCourse = (clientCourse: IClientCoursePopulated): IEmployeeClientCourse => {
   const { status, progress, date } = clientCourse;
@@ -17,42 +16,4 @@ const mapEmployeeClientCourse = (clientCourse: IClientCoursePopulated): IEmploye
   return emplooyeeClientCourse;
 };
 
-const mapEmployeeStack = (stack: TUserStackMember[]): TUserStackMemberShort[] =>
-  stack.map((stackMember) => ({
-    isPrimary: stackMember.isPrimary,
-    member: { name: stackMember.member.name },
-  }));
-
-const mapEmployeeInfo = (employee: IUser, courses: IClientCoursePopulated[]): IEmployeeInfo => ({
-  _id: employee._id,
-  firstName: employee.firstName,
-  lastName: employee.lastName,
-  position: employee.position,
-  technologies: employee.technologies,
-  group: employee.group,
-  avatar: employee.avatar,
-  birthday: employee.birthday,
-  skype: employee.skype,
-  phone: employee.phone,
-  rank: employee.rank,
-  courses: courses.map(mapEmployeeClientCourse),
-  stack: mapEmployeeStack(convertToTypeUnsafe<TUserStackMember[]>(employee.stack)),
-});
-
-const mapEmployeeShortInfo = (employee: IUser): IEmployeeShortInfo => ({
-  _id: employee._id,
-  firstName: employee.firstName,
-  lastName: employee.lastName,
-  position: employee.position,
-  group: employee.group,
-  avatar: employee.avatar,
-  skype: employee.skype,
-  phone: employee.phone,
-  rank: employee.rank,
-  stack: mapEmployeeStack(convertToTypeUnsafe<TUserStackMember[]>(employee.stack)),
-});
-
-const mapEmployeesShortInfo = (employees: IUser[]): IEmployeeShortInfo[] =>
-  employees.map((employee) => mapEmployeeShortInfo(employee));
-
-export { mapEmployeeInfo, mapEmployeesShortInfo };
+export { mapEmployeeStack, mapEmployeeClientCourse };
